Added tournaments authorization and participant dashboard

+++ app/controllers/host/tournaments_controller.rb
@@ -1,10 +1,27 @@
+
 class Host::TournamentsController < ApplicationController
-  before_action :authenticate_user!, except: [ :index, :show ]
+  before_action :authenticate_user!, except: [ :index, :show, :count ]
   before_action :set_tournament, only: [ :show, :edit, :update, :destroy ]
-  # before_action :authorize_tournament, only: [:new, :create]
+  before_action :authorize_tournament, only: [ :new, :create ]
+  before_action :authorize_existing_tournament, only: [ :edit, :update, :destroy ]
+
+  include Pundit
+
+  def count
+    authorize Tournament # Ensure public access
+    total_tournaments = Tournament.count
+    render json: { total_tournaments: total_tournaments }
+  end
+
+  # def size
+  #   authorize Tournament # Ensure public access
+  #   total_tournaments = Tournament.count
+  #   render json: { total_tournaments: total_tournaments }
+  # end
 
   def index
-    @tournaments = Tournament.all
+    @tournaments = Tournament.order(created_at: :desc) # Newest first
+    authorize Tournament
 
     respond_to do |format|
       format.html { render :index }
@@ -13,6 +30,8 @@ class Host::TournamentsController < ApplicationController
   end
 
   def show
+    authorize @tournament
+
     respond_to do |format|
       format.html { render :show }
       format.json { render json: @tournament }
@@ -21,12 +40,12 @@ class Host::TournamentsController < ApplicationController
 
   def new
     @tournament = Tournament.new
-    # authorize @tournament
+    authorize @tournament
   end
 
   def create
     @tournament = current_user.hosted_tournaments.build(tournament_params)
-    # authorize @tournament
+    authorize @tournament
 
     if @tournament.save
       respond_to do |format|
@@ -41,12 +60,42 @@ class Host::TournamentsController < ApplicationController
     end
   end
 
+  # def edit
+  #   authorize @tournament
+  # end
+
+  # def update
+  #   authorize @tournament
+  #   if @tournament.update(tournament_params)
+  #     respond_to do |format|
+  #       format.html { redirect_to host_tournament_path(@tournament), notice: "Tournament updated successfully!" }
+  #       format.json { render json: @tournament }
+  #     end
+  #   else
+  #     respond_to do |format|
+  #       format.html { render :edit }
+  #       format.json { render json: { errors: @tournament.errors.full_messages }, status: :unprocessable_entity }
+  #     end
+  #   end
+  # end
+
+
   def edit
-    # authorize @tournament
+    @tournament = Tournament.find(params[:id])
+    authorize @tournament # Ensure user is authorized to edit
+  rescue Pundit::NotAuthorizedError
+    redirect_to root_path, alert: "You are not authorized to edit this tournament."
   end
 
   def update
-    # authorize @tournament
+    @tournament = Tournament.find_by(id: params[:id])
+
+    unless @tournament
+      redirect_to host_tournaments_path, alert: "Tournament not found!" and return
+    end
+
+    authorize @tournament
+
     if @tournament.update(tournament_params)
       respond_to do |format|
         format.html { redirect_to host_tournament_path(@tournament), notice: "Tournament updated successfully!" }
@@ -54,14 +103,19 @@ class Host::TournamentsController < ApplicationController
       end
     else
       respond_to do |format|
-        format.html { render :edit }
+        format.html { render :edit, status: :unprocessable_entity }
         format.json { render json: { errors: @tournament.errors.full_messages }, status: :unprocessable_entity }
       end
     end
+  rescue Pundit::NotAuthorizedError
+    redirect_to root_path, alert: "You are not authorized to update this tournament."
   end
 
+
+
   def destroy
-    # authorize @tournament
+    authorize @tournament
+
     if @tournament.destroy
       respond_to do |format|
         format.html { redirect_to host_tournaments_path, notice: "Tournament deleted successfully!" }
@@ -86,6 +140,10 @@ class Host::TournamentsController < ApplicationController
   end
 
   def authorize_tournament
-    authorize Tournament.new
+    authorize Tournament
+  end
+
+  def authorize_existing_tournament
+    authorize @tournament
   end
 end
